Moved wpscan back to the main wptool file.
It's going to be too hard to maintain it separately,
and having to load yet another file is going to be a
pain, and they may not know that they need to. Just
better to keep it in the same file. I may take the strings
and separate them into another file that can be loaded
when needed, like how wpver gets a list of database
version numbers.

Signed-off-by: Matt Bagley <firstlife22@gmail.com>